refatorando os controllers

+++ app/controllers/events.py
@@ -16,28 +16,48 @@ def list():
 
 @bp_events.route('/<int:id>', methods=['delete'])
 def delete(id):
-    Event.query.filter(Event.id == id).delete()
-    current_app.db.session.commit()
-    return jsonify('Deletado!!!!')
+    event_exists = Event.query.filter(Event.id == id).first()
+    if event_exists: 
+        Event.query.filter(Event.id == id).delete()
+        current_app.db.session.commit()
+        return jsonify('Deletado!!!!')
+    else:
+        return jsonify({
+            'message': 'Registro não encontrado!.'
+        }), 400
 
-
+    
 @bp_events.route('/<int:id>', methods=['update'])
 def update(identificador):
-    bs = EventSchema()
-    query = Event.query.filter(Event.id == id)
-    query.update(request.json)
-    current_app.db.session.commit()
-    return bs.jsonify(query.first())
+    event_exists = Event.query.filter(Event.id == id).first()
+    if event_exists: 
+        event_schema = EventSchema()
+        query = Event.query.filter(Event.id == id)
+        query.update(request.json)
+        current_app.db.session.commit()
+        return event_schema.jsonify(query.first())
+    else:
+        return jsonify({
+            'message': 'Registro não encontrado!.'
+        }), 400
 
 
 @bp_events.route('/', methods=['post'])
 def create():
-    es = EventSchema()
-    event, error = es.load(request.json)
+    event_schema = EventSchema()    
+    username = request.json["username"]
+    password = request.json["password"]
+    
+    event_exists = Event.query.filter_by(username=username).first()
 
-    if error:
-        return jsonify(error), 401
+    if not event_exists:
+        new_event = Event(username, password)
+    else:
+        return jsonify({
+            'message': 'Este evento já foi registrado!.'
+        }), 400
 
-    current_app.db.session.add(event)
+    current_app.db.session.add(new_event)
     current_app.db.session.commit()
-    return es.jsonify(event), 201
+
+    return event_schema.jsonify(new_event), 201
